% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbb_pvalue.R
\name{bbb_fast_pvalue}
\alias{bbb_fast_pvalue}
\title{Description here..}
\usage{
bbb_fast_pvalue(n1_tot, n2_tot, gamma, n1_suc, n2_suc, p_slot, alpha)
}
\arguments{
\item{gamma}{...}

\item{n_tot}{...}

\item{n_suc}{...}
}
\description{
Description here..
}
\details{
...
}
