% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bscCI.R
\name{bscCI}
\alias{bscCI}
\title{Blyth-Still-Casella confidence interval}
\usage{
bscCI(n_tot, n_suc, conf)
}
\arguments{
\item{n_tot}{Total number of experiments}

\item{n_suc}{Number of successes}

\item{conf}{Confidence level (1-alpha)}
}
\description{
Blyth-Still-Casella confidence interval
}
\details{
...
}
