% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbscCI-package.R
\docType{package}
\name{rbscCI}
\alias{rbscCI}
\alias{rbscCI-package}
\title{Blyth-Still-Casella Confidence Interval}
\description{
Blyth-Still-Casella Confidence Interval
}
\details{
Provides a fast calculation of the Blyth-Still-Casella confidence interval.
}
