% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bscCI.R
\name{bscCI}
\alias{bscCI}
\title{Blyth-Still-Casella confidence interval}
\usage{
bscCI(n_tot, n_suc, conf, digits = 2)
}
\arguments{
\item{n_tot}{Total number of experiments}

\item{n_suc}{Number of successes}

\item{conf}{Confidence level (1-alpha)}

\item{digits}{Number of decimal places to be used}
}
\value{
A vector containing the confidence interval. If \code{digits} is given, both upper and lower limits are rounded to the given number of digits.
}
\description{
Blyth-Still-Casella confidence interval
}
\details{
Computes the exact Blyth-Still-Casella binomial confidence interval. The initial CI is the Clopper-Pearson confidence interval.
}
\examples{
bscCI(100,25,0.95,digits = 3)

}
