% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bscCI-package.R, R/bscCI.R
\docType{package}
\name{bscCI}
\alias{bscCI}
\alias{bscCI-package}
\alias{bscCI}
\title{Blyth-Still-Casella confidence intervals}
\usage{
bscCI(n_tot, n_suc, conf)
}
\arguments{
\item{n_tot}{...}

\item{n_suc}{...}

\item{conf}{...}
}
\description{
Blyth-Still-Casella confidence intervals

Description here..
}
\details{
Some details here.

...
}
