% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpCI.R
\name{cpCI}
\alias{cpCI}
\title{Clopper-Pearson confidence interval}
\usage{
cpCI(n_tot, n_suc, conf, digits = 2)
}
\arguments{
\item{n_tot}{Total number of experiments}

\item{n_suc}{Number of successes}

\item{conf}{Confidence level (1-alpha)}

\item{digits}{Number of decimal places to be used}
}
\description{
Clopper-Pearson confidence interval
}
\details{
Computes the Clopper-Pearson confidence interval.
}
\examples{
cpCI(100,25,0.95)

}
